Render queue: unsigned retry deadline, thumbnail visibility per tick
The deadline was computed with signed addition, undefined when the
millisecond clock nears the top of its 32-bit range; the due check was
already unsigned. The test now runs the whole path: failure near the
limit, deadline, wait, retry after the wrap.

Thumbnail visibility came from the last draw, so a cell scrolled out of
the sidebar or hidden behind the Outline page kept its old flag and could
render or retry unseen. The tick now recomputes it from the cells' boxes
and the sidebar's active page.

+++ src/renderq.h
@@ -82,7 +82,8 @@ static void renderq_done(struct RenderQ *q, int ok, LONG now)
     else
     {
         q->state = RS_RETRY;
-        q->retry_at_ms = now + RENDER_RETRY_MS;
+        /* Unsigned, like the check in renderq_due(): the clock may wrap. */
+        q->retry_at_ms = (LONG)((unsigned int)now + RENDER_RETRY_MS);
     }
 }
